feat: automated migration
