+++ internal-enrichment/greynoise/__metadata__/connector_config_schema.json
@@ -1,0 +1,78 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://www.filigran.io/connectors/greynoise_config.schema.json",
+  "type": "object",
+  "properties": {
+    "OPENCTI_URL": {
+      "description": "The base URL of the OpenCTI instance.",
+      "format": "uri",
+      "maxLength": 2083,
+      "minLength": 1,
+      "type": "string"
+    },
+    "OPENCTI_TOKEN": {
+      "description": "The API token to connect to OpenCTI.",
+      "type": "string"
+    },
+    "CONNECTOR_NAME": {
+      "default": "Greynoise",
+      "description": "The name of the connector.",
+      "type": "string"
+    },
+    "CONNECTOR_SCOPE": {
+      "default": [],
+      "description": "The scope of the connector.",
+      "items": {
+        "type": "string"
+      },
+      "type": "array"
+    },
+    "CONNECTOR_LOG_LEVEL": {
+      "default": "error",
+      "description": "The minimum level of logs to display.",
+      "enum": [
+        "debug",
+        "info",
+        "warn",
+        "warning",
+        "error"
+      ],
+      "type": "string"
+    },
+    "CONNECTOR_TYPE": {
+      "const": "INTERNAL_ENRICHMENT",
+      "default": "INTERNAL_ENRICHMENT",
+      "type": "string"
+    },
+    "CONNECTOR_AUTO": {
+      "default": false,
+      "description": "Whether the connector should run automatically when an entity is created or updated.",
+      "type": "boolean"
+    },
+    "GREYNOISE_KEY": {
+      "description": "The GreyNoise API key.",
+      "type": "string"
+    },
+    "GREYNOISE_MAX_TLP": {
+      "description": "Maximum TLP level for data to be sent to GreyNoise.",
+      "type": "string"
+    },
+    "GREYNOISE_SIGHTING_NOT_SEEN": {
+      "description": "Create sighting with count=0 when IP not seen.",
+      "type": "boolean"
+    },
+    "GREYNOISE_NO_SIGHTINGS": {
+      "description": "Skip any sighting creations.",
+      "type": "boolean"
+    }
+  },
+  "required": [
+    "OPENCTI_URL",
+    "OPENCTI_TOKEN",
+    "GREYNOISE_KEY",
+    "GREYNOISE_MAX_TLP",
+    "GREYNOISE_SIGHTING_NOT_SEEN",
+    "GREYNOISE_NO_SIGHTINGS"
+  ],
+  "additionalProperties": true
+}
