+++ internal-enrichment/greynoise/__metadata__/connector_manifest.json
@@ -15,7 +15,7 @@
   "support_version": ">=5.6.1",
   "subscription_link": null,
   "source_code": "https://github.com/OpenCTI-Platform/connectors/tree/master/internal-enrichment/greynoise",
-  "manager_supported": false,
+  "manager_supported": true,
   "container_version": "rolling",
   "container_image": "opencti/connector-greynoise",
   "container_type": "INTERNAL_ENRICHMENT"
