+++ internal-enrichment/greynoise/src/connector/__init__.py
@@ -1,0 +1,7 @@
+from connector.connector import GreyNoiseConnector
+from connector.settings import ConnectorSettings
+
+__all__ = [
+    "GreyNoiseConnector",
+    "ConnectorSettings",
+]
