+++ internal-enrichment/greynoise/src/main.py
@@ -1,956 +1,26 @@
-import os
-from datetime import datetime, timedelta
-from typing import Dict
+import traceback
 
-import pycountry
-import stix2
-import yaml
-from dateutil.parser import parse
-from greynoise.api import APIConfig, GreyNoise
-from greynoise.exceptions import RequestFailure
-from pycti import (
-    Identity,
-    Indicator,
-    Location,
-    Malware,
-    OpenCTIConnectorHelper,
-    StixCoreRelationship,
-    StixSightingRelationship,
-    ThreatActorGroup,
-    Tool,
-    Vulnerability,
-    get_config_variable,
-)
-
-INTEGRATION_NAME = "opencti-enricher-v4.0"
-
-
-class GreyNoiseConnector:
-    def __init__(self):
-        # Instantiate the connector helper from config
-        config_file_path = os.path.dirname(os.path.abspath(__file__)) + "/config.yml"
-        config = (
-            yaml.load(open(config_file_path), Loader=yaml.FullLoader)
-            if os.path.isfile(config_file_path)
-            else {}
-        )
-        self.helper = OpenCTIConnectorHelper(config, playbook_compatible=True)
-        self.greynoise_key = get_config_variable(
-            "GREYNOISE_KEY", ["greynoise", "key"], config
-        )
-        self.max_tlp = get_config_variable(
-            "GREYNOISE_MAX_TLP", ["greynoise", "max_tlp"], config
-        )
-        self.sighting_not_seen = get_config_variable(
-            "GREYNOISE_SIGHTING_NOT_SEEN", ["greynoise", "sighting_not_seen"], config
-        )
-        self.no_sightings = get_config_variable(
-            "GREYNOISE_NO_SIGHTINGS",
-            ["greynoise", "no_sightings"],
-            config,
-            default=False,
-        )
-        self.greynoise_ent_name = get_config_variable(
-            "GREYNOISE_NAME",
-            ["greynoise", "name"],
-            config,
-            default="GreyNoise Intelligence",
-        )
-        self.greynoise_ent_desc = get_config_variable(
-            "GREYNOISE_DESCRIPTION",
-            ["greynoise", "description"],
-            config,
-            default="GreyNoise collects and analyzes untargeted, widespread, and opportunistic scan and attack activity that reaches every server directly connected to the Internet.",
-        )
-        self.indicator_score_malicious = get_config_variable(
-            "GREYNOISE_INDICATOR_SCORE_MALICIOUS",
-            ["greynoisefeed", "indicator_score_malicious"],
-            config,
-            isNumber=True,
-            default=75,
-        )
-        self.indicator_score_suspicious = get_config_variable(
-            "GREYNOISE_INDICATOR_SCORE_SUSPICIOUS",
-            ["greynoisefeed", "indicator_score_suspicious"],
-            config,
-            isNumber=True,
-            default=50,
-        )
-        self.indicator_score_benign = get_config_variable(
-            "GREYNOISE_INDICATOR_SCORE_BENIGN",
-            ["greynoisefeed", "indicator_score_benign"],
-            config,
-            isNumber=True,
-            default=20,
-        )
-
-        # Define variables
-        self._CONNECTOR_RUN_INTERVAL_SEC = 60 * 60
-        self.tlp = None
-        self.stix_objects = []
-
-    def _get_indicator_score(self, classification):
-        if classification == "malicious":
-            self.indicator_score = self.indicator_score_malicious
-        elif classification == "suspicious" or classification == "2":
-            self.indicator_score = self.indicator_score_suspicious
-        else:
-            self.indicator_score = self.indicator_score_benign
-
-        return self.indicator_score
-
-    def _extract_and_check_markings(self, opencti_entity: dict) -> bool:
-        """
-        Extract TLP, and we check if the variable "max_tlp" is less than
-        or equal to the markings access of the entity from OpenCTI.
-        If this is true, we can send the data to connector for enrichment.
-
-        :param opencti_entity: Parameter that contains all information about the entity,
-                               including "objectMarking", the marking that the observable uses.
-        :return: A boolean
-        """
-
-        for marking_definition in opencti_entity["objectMarking"]:
-            if marking_definition["definition_type"] == "TLP":
-                self.tlp = marking_definition["definition"]
-
-        is_valid_max_tlp = OpenCTIConnectorHelper.check_max_tlp(self.tlp, self.max_tlp)
-
-        return is_valid_max_tlp
-
-    def _generate_stix_relationship(
-        self,
-        source_ref: str,
-        stix_core_relationship_type: str,
-        target_ref: str,
-        start_time: str | None = None,
-        stop_time: str | None = None,
-    ) -> dict:
-        """
-        This method allows you to create a relationship in Stix2 format.
-
-        :param source_ref: This parameter is the "from" of the relationship.
-        :param stix_core_relationship_type: Parameter,
-        :param target_ref: This parameter is the "to" of the relationship.
-        :param start_time: This parameter is the start of the relationship. Value not required, None by default.
-        :param stop_time: This parameter is the stop of the relationship. Value not required, None by default.
-        :return: A dict
-        """
-
-        return stix2.Relationship(
-            id=StixCoreRelationship.generate_id(
-                stix_core_relationship_type, source_ref, target_ref
-            ),
-            relationship_type=stix_core_relationship_type,
-            source_ref=source_ref,
-            start_time=start_time,
-            stop_time=stop_time,
-            target_ref=target_ref,
-            created_by_ref=self.greynoise_identity["id"],
-        )
-
-    def _create_custom_label(self, name_label: str, color_label: str):
-        """
-        This method allows you to create a custom label, using the OpenCTI API.
-
-        :param name_label: A parameter giving the name of the label.
-        :param color_label: A parameter giving the color of the label.
-        """
-
-        new_custom_label = self.helper.api.label.read_or_create_unchecked(
-            value=name_label, color=color_label, createdBy=self.greynoise_identity["id"]
-        )
-        if new_custom_label is None:
-            self.helper.connector_logger.error(
-                "[ERROR] The label could not be created. If your connector does not have the permission to create labels, "
-                "please create it manually before launching",
-                {"name_label": name_label},
-            )
-        else:
-            self.all_labels.append(new_custom_label["value"])
-
-    def _process_labels(self, data: dict) -> tuple:
-        """
-        This method allows you to start the process of creating labels and recovering associated malware.
-
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        :param data_tags: A parameter that contains all the data relating to the existing tags in GreyNoise
-        :return: A tuple (all labels, all malwares)
-        """
-
-        self.all_labels = []
-        all_malwares = []
-        entity_tags = data["internet_scanner_intelligence"].get("tags", [])
-
-        if data["internet_scanner_intelligence"]["classification"] == "benign":
-            # Create label GreyNoise "benign"
-            self._create_custom_label("gn-classification: benign", "#06c93a")
-            # Include additional label "benign-actor"
-            self._create_custom_label(
-                f"gn-benign-actor: {data['internet_scanner_intelligence']['actor']} ",
-                "#06c93a",
-            )
-        elif data["internet_scanner_intelligence"]["classification"] == "unknown":
-            # Create label GreyNoise "unknown"
-            self._create_custom_label("gn-classification: unknown", "#a6a09f")
-        elif data["internet_scanner_intelligence"]["classification"] == "malicious":
-            # Create label GreyNoise "malicious"
-            self._create_custom_label("gn-classification: malicious", "#ff8178")
-        elif data["internet_scanner_intelligence"]["classification"] == "suspicious":
-            # Create label GreyNoise "suspicious"
-            self._create_custom_label("gn-classification: suspicious", "#e3d922")
-
-        if data["business_service_intelligence"]["trust_level"] == "1":
-            # Create label GreyNoise "trust level 1"
-            self._create_custom_label("gn-trust-level: reasonably ignore", "#90D5FF")
-            # Include additional label "provider"
-            self._create_custom_label(
-                f"gn-provider: {data['business_service_intelligence']['name']} ",
-                "#90D5FF",
-            )
-        elif data["business_service_intelligence"]["trust_level"] == "2":
-            # Create label GreyNoise "trust level 1"
-            self._create_custom_label("gn-trust-level: commonly seen", "#57B9FF")
-            # Include additional label "provider"
-            self._create_custom_label(
-                f"gn-provider: {data['business_service_intelligence']['name']} ",
-                "#57B9FF",
-            )
-
-        if data["internet_scanner_intelligence"].get("bot") is True:
-            # Create label for "Known Bot Activity"
-            self._create_custom_label("Known BOT Activity", "#7e4ec2")
-
-        if data["internet_scanner_intelligence"]["tor"] is True:
-            # Create label for "Known Tor Exit Node"
-            self._create_custom_label("Known TOR Exit Node", "#7e4ec2")
-
-        if data["internet_scanner_intelligence"]["vpn"] is True:
-            # Create label for "Known VPN"
-            self._create_custom_label("Known VPN", "#7e4ec2")
-
-        # Create all Labels in entity_tags
-        for tag in entity_tags:
-            # Create red label when malicious intent and type not category worm and activity
-            if tag["intention"] == "malicious" and tag["category"] not in ["worm"]:
-                self._create_custom_label(f"{tag['name']}", "#ff8178")
-
-            # If category is worm, prepare malware object
-            elif tag["category"] == "worm":
-                malware_worm = {
-                    "name": f"{tag}",
-                    "description": f"{tag['description']}",
-                    "type": "worm",
-                }
-                all_malwares.append(malware_worm)
-                self.all_labels.append(tag["name"])
-
-            elif tag["intention"] == "benign":
-                # Create green label for benign tags
-                self._create_custom_label(f"{tag['name']}", "#06c93a")
-            elif tag["intention"] == "suspicious":
-                # Create yellow label for suspicious tags
-                self._create_custom_label(f"{tag['name']}", "#e3d922")
-            else:
-                # Create white label otherwise
-                self._create_custom_label(f"{tag['name']}", "#ffffff")
-
-        return self.all_labels, all_malwares
-
-    def _generate_stix_external_reference(self, data: dict) -> list:
-        """
-        This method allows you to create an external reference in Stix2 format.
-
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        :return: list -> ExternalReference (Stix2 format)
-        """
-        description = ""
-        if data["internet_scanner_intelligence"]["found"] is True:
-            description = "This reference will direct to the GreyNoise Visualizer IP details page for an IP that has been seen mass scanning the internet."
-        elif data["business_service_intelligence"]["found"] is True:
-            description = "This reference will direct to the GreyNoise Visualizer IP details page for an IP that is part of a common business service."
-        else:
-            description = "This reference will direct to the GreyNoise Visualizer IP details page for an IP that has not yet been identified by GreyNoise, meaning it has not been seen mass scanning the internet nor does it belong to a business service that we monitor."
-
-        # Generate ExternalReference
-        external_reference = stix2.ExternalReference(
-            source_name=self.greynoise_ent_name,
-            url=f"https://viz.greynoise.io/ip/{data['ip']}",
-            external_id=data["ip"],
-            description=description,
-        )
-        return [external_reference]
-
-    def _generate_greynoise_stix_identity(self):
-        """
-        This method create and adds in "self.stix_objects" the "Identity (organization)" of GreyNoise in Stix2 format,
-        its full name is configurable in the environment variable.
-        """
-
-        # Generate "GreyNoise Sensor" Identity
-        self.greynoise_identity = stix2.Identity(
-            id=Identity.generate_id(self.greynoise_ent_name, "organization"),
-            name=self.greynoise_ent_name,
-            description=f"Connector Enrichment {self.greynoise_ent_name}",
-            identity_class="organization",
-        )
-        self.stix_objects.append(self.greynoise_identity)
-
-    def _generate_other_stix_identity_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Identity (organization)"
-        provided by GreyNoise in Stix2 format,
-
-        - Relationship : Observable -> "related-to" -> Organization.
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        organization = data["internet_scanner_intelligence"]["metadata"].get(
-            "organization", ""
-        )
-
-        if organization != "":
-            # Generate other Identity
-            stix_organization = stix2.Identity(
-                id=Identity.generate_id(organization, "organization"),
-                name=organization,
-                identity_class="organization",
-                created_by_ref=self.greynoise_identity["id"],
-            )
-            self.stix_objects.append(stix_organization)
-
-            # Generate Relationship : Observable -> "related-to" -> Organization
-            observable_to_organization = self._generate_stix_relationship(
-                self.stix_entity["id"], "related-to", stix_organization.id
-            )
-            self.stix_objects.append(observable_to_organization)
-
-    def _generate_stix_asn_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Autonomous System Number"
-        provided by GreyNoise in Stix2 format,
-
-        - Relationship : observable -> "belongs-to" -> Asn.
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        # Generate Asn
-        entity_asn = data["internet_scanner_intelligence"]["metadata"].get("asn", "")
-        if entity_asn != "":
-            asn_number = int(
-                data["internet_scanner_intelligence"]["metadata"]["asn"].replace(
-                    "AS", ""
-                )
-            )
-            stix_asn = stix2.AutonomousSystem(
-                type="autonomous-system",
-                number=asn_number,
-                name=entity_asn,
-                custom_properties={
-                    "created_by_ref": self.greynoise_identity["id"],
-                    "x_opencti_score": self.indicator_score,
-                },
-            )
-            self.stix_objects.append(stix_asn)
-
-            # Generate Relationship : observable -> "belongs-to" -> Asn
-            observable_to_asn = self._generate_stix_relationship(
-                self.stix_entity["id"], "belongs-to", stix_asn.id
-            )
-            self.stix_objects.append(observable_to_asn)
-
-    def _generate_stix_domain_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Domain"
-        provided by GreyNoise in Stix2 format,
-
-        - Relationship : observable -> "related-to" -> Domain.
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        # Generate Asn
-        entity_domain = data["internet_scanner_intelligence"]["metadata"].get(
-            "rdns", ""
-        )
-        if entity_domain != "":
-            stix_domain = stix2.DomainName(
-                type="domain-name",
-                value=entity_domain,
-                custom_properties={
-                    "created_by_ref": self.greynoise_identity["id"],
-                    "x_opencti_score": self.indicator_score,
-                },
-            )
-            self.stix_objects.append(stix_domain)
-
-            # Generate Relationship : observable -> "resolves-to" -> Domain
-            domain_to_observable = self._generate_stix_relationship(
-                stix_domain.id, "resolves-to", self.stix_entity["id"]
-            )
-            self.stix_objects.append(domain_to_observable)
-
-    def _generate_stix_location_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Location (City + Country)"
-        provided by GreyNoise in Stix2 format.
-
-        - Relationships : (observable -> "located-at" -> city) and (city -> "located-at" -> country)
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        if data["internet_scanner_intelligence"]["found"] is True:
-            country = pycountry.countries.get(
-                alpha_2=data["internet_scanner_intelligence"]["metadata"][
-                    "source_country_code"
-                ]
-            )
-            country_name = (
-                country.official_name
-                if hasattr(country, "official_name")
-                else country.name
-            )
-
-            # Generate City Location
-            stix_city_location = stix2.Location(
-                id=Location.generate_id(
-                    data["internet_scanner_intelligence"]["metadata"]["source_city"],
-                    "City",
-                ),
-                name=data["internet_scanner_intelligence"]["metadata"]["source_city"],
-                country=country_name,
-                custom_properties={"x_opencti_location_type": "City"},
-            )
-            self.stix_objects.append(stix_city_location)
-
-            # Generate Relationship : observable -> "located-at" -> city
-            observable_to_city = self._generate_stix_relationship(
-                self.stix_entity["id"], "located-at", stix_city_location.id
-            )
-            self.stix_objects.append(observable_to_city)
-
-            # Generate Country Location
-            stix_country_location = stix2.Location(
-                id=Location.generate_id(country.name, "Country"),
-                name=country_name,
-                country=country_name,
-                custom_properties={
-                    "x_opencti_location_type": "Country",
-                    "x_opencti_aliases": [
-                        data["internet_scanner_intelligence"]["metadata"][
-                            "source_country_code"
-                        ]
-                    ],
-                },
-            )
-            self.stix_objects.append(stix_country_location)
-
-            # Generate Relationship : city -> "located-at" -> country
-            city_to_country = self._generate_stix_relationship(
-                stix_city_location.id, "located-at", stix_country_location.id
-            )
-            self.stix_objects.append(city_to_country)
-
-    def _generate_stix_vulnerability_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "vulnerability"
-        provided by GreyNoise in Stix2 format.
-
-        - Relationship : observable -> "related-to" -> vulnerability
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        if (
-            "cves" in data["internet_scanner_intelligence"]
-            and data["internet_scanner_intelligence"].get("cves", []) is not []
-        ):
-            entity_vulns = data["internet_scanner_intelligence"]["cves"]
-            for vuln in entity_vulns:
-                # Generate Vulnerability
-                stix_vulnerability = stix2.Vulnerability(
-                    id=Vulnerability.generate_id(vuln),
-                    name=vuln,
-                    created_by_ref=self.greynoise_identity["id"],
-                    allow_custom=True,
-                )
-                self.stix_objects.append(stix_vulnerability)
-
-                # Generate Relationship : observable -> "related-to" -> vulnerability
-                observable_to_vulnerability = self._generate_stix_relationship(
-                    self.stix_entity["id"],
-                    "related-to",
-                    stix_vulnerability.id,
-                    self.first_seen,
-                    self.last_seen,
-                )
-                self.stix_objects.append(observable_to_vulnerability)
-
-    def _generate_stix_tool_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "tool"
-        provided by GreyNoise in Stix2 format.
-
-        - Relationship : observable -> "related-to" -> tool
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        if data["internet_scanner_intelligence"]["vpn"] is True:
-            # Generate Tool
-            stix_tool = stix2.Tool(
-                id=Tool.generate_id(
-                    f"VPN: {data['internet_scanner_intelligence']['vpn_service']}"
-                ),
-                name=f"VPN: {data['internet_scanner_intelligence']['vpn_service']}",
-                labels=["tool"],
-                created_by_ref=self.greynoise_identity["id"],
-                custom_properties={
-                    "x_opencti_aliases": data["internet_scanner_intelligence"][
-                        "vpn_service"
-                    ]
-                },
-                allow_custom=True,
-            )
-            self.stix_objects.append(stix_tool)
-
-            # Generate Relationship : observable -> "related-to" -> tool
-            observable_to_tool = self._generate_stix_relationship(
-                self.stix_entity["id"],
-                "related-to",
-                stix_tool.id,
-                self.first_seen,
-                self.last_seen,
-            )
-            self.stix_objects.append(observable_to_tool)
-
-    def _generate_stix_sighting(
-        self,
-        external_reference: list,
-        stix_indicator: dict,
-        sighting_not_seen: bool = False,
-    ):
-        """
-        This method creates a sighting.
-
-        - If the IPv4 is known by GreyNoise:
-
-          - Create a `sighting` from the IPv4 observable to the GreyNoise entity with `count=1`
-
-        - If the IPv4 is not knew by GreyNoise:
-
-          - if `GREYNOISE_SIGHTING_NOT_SEEN=true`: create a `sighting` from the IPv4 observable to the
-            GreyNoise entity with `count=0`
-          - if `GREYNOISE_SIGHTING_NOT_SEEN=false`: do nothing.
-
-        :param external_reference: This parameter contains the list external reference associated with the IPv4.
-        :param sighting_not_seen: This parameter is a boolean that corresponds to the configuration of
-                                  sighting_not_seen, which aims to know if the user wants to create a sighting at 0.
-        """
-
-        default_now = datetime.now().strftime("%Y-%m-%dT%H:%M:%SZ")
-        # Create sighting for target entity
-        first_seen = default_now if sighting_not_seen is True else self.first_seen
-        last_seen = default_now if sighting_not_seen is True else self.last_seen
-        sighting_count = 0 if sighting_not_seen is True else 1
-        stix_sighting_entity = stix2.Sighting(
-            id=StixSightingRelationship.generate_id(
-                self.stix_entity["id"],
-                self.greynoise_identity["id"],
-                first_seen,
-                last_seen,
-            ),
-            first_seen=first_seen,
-            last_seen=last_seen,
-            count=sighting_count,
-            description=self.greynoise_ent_desc,
-            created_by_ref=self.greynoise_identity["id"],
-            where_sighted_refs=[self.greynoise_identity["id"]],
-            external_references=external_reference,
-            object_marking_refs=stix2.TLP_WHITE,
-            # As SDO are not supported in official STIX, we use a fake ID in ref
-            # Worker will use custom_properties instead
-            sighting_of_ref="indicator--51b92778-cef0-4a90-b7ec-ebd620d01ac8",  # Fake
-            custom_properties={
-                "x_opencti_sighting_of_ref": self.stix_entity["id"],
-                "x_opencti_negative": True,
-            },
-        )
-        self.stix_objects.append(stix_sighting_entity)
-        # Create sighting for associated indicator
-        stix_sighting_indicator = stix2.Sighting(
-            id=StixSightingRelationship.generate_id(
-                stix_indicator["id"],
-                self.greynoise_identity["id"],
-                first_seen,
-                last_seen,
-            ),
-            first_seen=first_seen,
-            last_seen=last_seen,
-            count=sighting_count,
-            description=self.greynoise_ent_desc,
-            created_by_ref=self.greynoise_identity["id"],
-            sighting_of_ref=stix_indicator["id"],
-            where_sighted_refs=[self.greynoise_identity["id"]],
-            external_references=external_reference,
-            object_marking_refs=stix2.TLP_WHITE,
-            custom_properties={
-                "x_opencti_negative": True,
-            },
-        )
-        self.stix_objects.append(stix_sighting_indicator)
-
-    def _generate_stix_malware_with_relationship(self, malwares: list):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "malware"
-        provided by GreyNoise in Stix2 format.
-
-        - Relationship : observable -> "related-to" -> malware
-        :param malwares: This parameter contains a list of all malwares associated with the IPv4 of the searched
-                         observable, whose malwares is retrieved when comparing labels and data_tags.
-        """
-
-        for malware in malwares:
-            stix_malware = stix2.Malware(
-                id=Malware.generate_id(malware["name"]),
-                created_by_ref=self.greynoise_identity["id"],
-                name=malware["name"],
-                description=malware["description"],
-                is_family=False,
-                malware_types=malware["type"] if malware["type"] == "worm" else None,
-                created=self.first_seen,
-            )
-            self.stix_objects.append(stix_malware)
-
-            # Generate Relationship : observable -> "related-to" -> malware
-            observable_to_malware = self._generate_stix_relationship(
-                self.stix_entity["id"],
-                "related-to",
-                stix_malware.id,
-                self.first_seen,
-                self.last_seen,
-            )
-            self.stix_objects.append(observable_to_malware)
-
-    def _generate_stix_threat_actor_with_relationship(self, data: dict):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Threat Actor"
-        provided by GreyNoise in Stix2 format.
-
-        - Relationship : observable -> "related-to" -> threat actor
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        """
-
-        # Create threat actor for non-benign when known
-        if (
-            data["internet_scanner_intelligence"]["actor"]
-            and data["internet_scanner_intelligence"]["actor"] != "unknown"
-            and data["internet_scanner_intelligence"]["actor"] != ""
-            and data["internet_scanner_intelligence"]["classification"] != "benign"
-        ):
-            # Generate Threat Actor
-            stix_threat_actor = stix2.ThreatActor(
-                id=ThreatActorGroup.generate_id(
-                    data["internet_scanner_intelligence"]["actor"]
-                ),
-                name=data["internet_scanner_intelligence"]["actor"],
-                created_by_ref=self.greynoise_identity["id"],
-            )
-            self.stix_objects.append(stix_threat_actor)
-
-            # Generate Relationship : observable -> "related-to" -> threat actor
-            observable_to_threat_actor = self._generate_stix_relationship(
-                self.stix_entity["id"],
-                "related-to",
-                stix_threat_actor.id,
-                self.first_seen,
-                self.last_seen,
-            )
-            self.stix_objects.append(observable_to_threat_actor)
-
-    def _generate_stix_indicator_with_relationship(
-        self, data: dict, detection: bool, external_reference: list, labels: list = None
-    ) -> dict:
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Indicator"
-        in Stix2 format.
-
-        - Relationship : Indicator -> "based-on" -> Observable
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        :param detection: If sighting_not_seen is true, then this detection parameter is false and vice versa.
-        :param external_reference: This parameter contains the list external reference associated with the IPv4.
-        :param labels: This parameter contains a list of all labels associated with the IPv4.
-        :return: dict
-        """
-
-        if self.sighting_not_seen is True:
-            self.indicator_score = 0
-
-        description = (
-            "Internet Scanning IP detected by GreyNoise with classification `"
-            + data["internet_scanner_intelligence"]["classification"]
-            + "`."
-        )
-
-        # Generate Indicator
-        stix_indicator = stix2.Indicator(
-            id=Indicator.generate_id(data["ip"]),
-            name=data["ip"],
-            labels=labels if detection is True else [],
-            pattern=f"[ipv4-addr:value = '{data['ip']}']",
-            created_by_ref=self.greynoise_identity["id"],
-            external_references=external_reference,
-            description=description,
-            custom_properties={
-                "pattern_type": "stix",
-                "x_opencti_score": self.indicator_score,
-                "x_opencti_main_observable_type": "IPv4-Addr",
-                "detection": True if detection is True else False,
-            },
-        )
-        self.stix_objects.append(stix_indicator)
-
-        # Generate Relationship : Indicator -> "based-on" -> Observable
-        indicator_to_observable = self._generate_stix_relationship(
-            stix_indicator.id, "based-on", self.stix_entity["id"]
-        )
-        self.stix_objects.append(indicator_to_observable)
-
-        return stix_indicator
-
-    def _generate_stix_observable(
-        self, data: dict, detection: bool, external_reference: list, labels: list = None
-    ):
-        """
-        This method creates and adds a bundle to "self.stix_objects" the IPv4 associated "Observable"
-        in Stix2 format.
-
-        :param detection: If sighting_not_seen is true, then this detection parameter is false and vice versa.
-        :param external_reference: This parameter contains the list external reference associated with the IPv4.
-        :param labels: This parameter contains a list of all labels associated with the IPv4.
-        """
-
-        description = (
-            "Internet Scanning IP detected by GreyNoise with classification `"
-            + data["internet_scanner_intelligence"]["classification"]
-            + "`."
-        )
-
-        # Generate Observable
-        stix_observable = stix2.IPv4Address(
-            id=self.stix_entity["id"],
-            type="ipv4-addr",
-            value=self.stix_entity["value"],
-            custom_properties={
-                "x_opencti_external_references": external_reference,
-                "x_opencti_score": self.indicator_score,
-                "x_opencti_labels": labels if detection is True else [],
-                "x_opencti_created_by_ref": self.greynoise_identity["id"],
-                "x_opencti_description": description,
-            },
-        )
-        self.stix_objects.append(stix_observable)
-
-    def _generate_stix_bundle(self, data: dict, stix_entity: dict) -> str:
-        """
-        This method create a bundle in Stix2 format.
-
-        :param data: A parameter that contains all the data about the IPv4 that was searched for in GreyNoise.
-        :param data_tags: A parameter that contains all the data relating to the existing tags in GreyNoise.
-        :param stix_entity: A parameter that contains all the IPv4 information in OpenCTI.
-        :return: str bundle
-        """
-
-        self.stix_entity = stix_entity
-        self._generate_greynoise_stix_identity()
-
-        if (
-            data["internet_scanner_intelligence"]["found"] is False
-            and data["business_service_intelligence"]["found"] is False
-            and self.sighting_not_seen is True
-        ):
-            """
-            If the IP has not been identified by GreyNoise, but the user still wants to create a sighting at count=0,
-            they can do so by setting the sighting_not_seen variable to "true", in this case we create an external
-            reference linked in the sighting.
-            """
-
-            self.helper.connector_logger.info(
-                "[CONNECTOR] The IPv4 has not been identified, but the creation of the sighting is true",
-                {
-                    "IPv4": stix_entity["value"],
-                    "config_sighting_not_seen": self.sighting_not_seen,
-                },
-            )
-
-            external_reference = self._generate_stix_external_reference(data)
-            stix_indicator = self._generate_stix_indicator_with_relationship(
-                data, False, external_reference
-            )
-            self.all_labels = []
-            self._create_custom_label("greynoise: not observed", "#ffffff")
-
-            self._generate_stix_observable(True, external_reference, self.all_labels)
-            if not self.no_sightings:
-                self._generate_stix_sighting(external_reference, stix_indicator, True)
-
-        else:
-            self.helper.connector_logger.info(
-                "[CONNECTOR] IPv4, has been identified by GreyNoise and generation of the Stix bundle is in progress.",
-                {"IPv4": stix_entity["value"]},
-            )
-
-            if data["internet_scanner_intelligence"]["found"] is True:
-                self.last_seen = parse(
-                    data["internet_scanner_intelligence"]["last_seen_timestamp"]
-                ).strftime("%Y-%m-%dT%H:%M:%SZ")
-                self._get_indicator_score(
-                    data["internet_scanner_intelligence"].get(
-                        "classification", "unknown"
-                    )
-                )
-            else:
-                self.last_seen = parse(
-                    data["business_service_intelligence"]["last_updated"]
-                ).strftime("%Y-%m-%dT%H:%M:%SZ")
-                self._get_indicator_score(
-                    data["business_service_intelligence"].get("trust_level", "2")
-                )
-
-            if (
-                "first_seen" in data["internet_scanner_intelligence"]
-                and data["internet_scanner_intelligence"]["first_seen"] != ""
-            ):
-                self.first_seen = parse(
-                    data["internet_scanner_intelligence"]["first_seen"]
-                ).strftime("%Y-%m-%dT%H:%M:%SZ")
-            else:
-                self.first_seen = self.last_seen
-                self.last_seen = datetime.fromisoformat(
-                    self.last_seen.replace("Z", "+00:00")
-                ) + timedelta(hours=23)
-                self.last_seen = self.last_seen.strftime("%Y-%m-%dT%H:%M:%SZ")
-
-            # Generate Stix Object for bundle
-            labels, malwares = self._process_labels(data)
-            external_reference = self._generate_stix_external_reference(data)
-            stix_indicator = self._generate_stix_indicator_with_relationship(
-                data, True, external_reference, labels
-            )
-
-            self._generate_other_stix_identity_with_relationship(data)
-            self._generate_stix_asn_with_relationship(data)
-            self._generate_stix_domain_with_relationship(data)
-            self._generate_stix_location_with_relationship(data)  # City + Country
-            self._generate_stix_vulnerability_with_relationship(data)
-            self._generate_stix_tool_with_relationship(data)
-            self._generate_stix_malware_with_relationship(malwares)
-            self._generate_stix_threat_actor_with_relationship(data)
-            self._generate_stix_observable(data, True, external_reference, labels)
-            if not self.no_sightings:
-                self._generate_stix_sighting(external_reference, stix_indicator, False)
-
-        uniq_bundles_objects = list(
-            {obj["id"]: obj for obj in self.stix_objects}.values()
-        )
-
-        self.helper.connector_logger.info(
-            "[CONNECTOR] For this Ipv4, the number of Stix bundle(s) that will be enriched.",
-            {
-                "IPv4": stix_entity["value"],
-                "Stix_bundle_length": len(uniq_bundles_objects),
-            },
-        )
-
-        stix2_bundle = self.helper.stix2_create_bundle(uniq_bundles_objects)
-        return stix2_bundle
-
-    def _process_message(self, data: Dict) -> str:
-        # Security to limit playbook triggers to something other than the scope initial
-        scopes = self.helper.connect_scope.lower().replace(" ", "").split(",")
-        entity_splited = data["entity_id"].split("--")
-        entity_type = entity_splited[0].lower()
-
-        if entity_type in scopes:
-            # OpenCTI entity information retrieval
-            stix_entity = data["stix_entity"]
-            opencti_entity = data["enrichment_entity"]
-            self.stix_objects = data["stix_objects"]
-
-            is_valid_max_tlp = self._extract_and_check_markings(opencti_entity)
-            if not is_valid_max_tlp:
-                raise ValueError(
-                    "[ERROR] Do not send any data, TLP of the observable is greater than MAX TLP, "
-                    "the connector does not has access to this observable, please check the group of the connector user"
-                )
-
-            # Extract Value from opencti entity data
-            opencti_entity_value = stix_entity["value"]
-
-            try:
-                # Get "IP Lookup" GreyNoise API Response
-                # https://docs.greynoise.io/reference/v3ip#/
-                api_config = APIConfig(
-                    api_key=self.greynoise_key, integration_name=INTEGRATION_NAME
-                )
-                session = GreyNoise(api_config)
-
-                json_data = session.ip(opencti_entity_value)
-
-                if (
-                    "found" in json_data["internet_scanner_intelligence"]
-                    and json_data["internet_scanner_intelligence"]["found"] is False
-                    and json_data["business_service_intelligence"]["found"] is False
-                    and self.sighting_not_seen is False
-                ):
-                    raise ValueError(
-                        "[API] This IP has not yet been identified by GreyNoise"
-                    )
-
-                # Generate a stix bundle
-                stix_bundle = self._generate_stix_bundle(json_data, stix_entity)
-
-                # Send stix2 bundle
-                bundles_sent = self.helper.send_stix2_bundle(stix_bundle)
-                return (
-                    "[CONNECTOR] Sent "
-                    + str(len(bundles_sent))
-                    + " stix bundle(s) for worker import"
-                )
-
-            except RequestFailure as e:
-                error_msg = str(e)
-                # Check if this is a rate limit (429) error
-                if "429" in error_msg or "too many" in error_msg.lower():
-                    self.helper.connector_logger.warning(
-                        "[RATE LIMIT] GreyNoise API rate limit exceeded, will retry later",
-                        {"ip": opencti_entity_value, "error": error_msg},
-                    )
-                    # Return a message instead of raising - allows retry
-                    return (
-                        "[RATE LIMIT] API rate limit exceeded, message will be retried"
-                    )
-                else:
-                    # For other API request failures, raise as before
-                    raise ValueError(
-                        "[ERROR] GreyNoise API Request Failed:",
-                        {"Exception": error_msg},
-                    )
-
-            except Exception as e:
-                # Handling other unexpected exceptions
-                raise ValueError(
-                    "[ERROR] Unexpected Error occurred:", {"Exception": str(e)}
-                )
-        else:
-            return self.helper.connector_logger.info(
-                "[INFO] The trigger does not concern the initial scope found in the config connector, "
-                "maybe choose a more specific filter in the playbook",
-                {"entity_id": data["entity_id"]},
-            )
-
-    # Start the main loop
-    def start(self):
-        self.helper.listen(message_callback=self._process_message)
-
+from connector import ConnectorSettings, GreyNoiseConnector
+from pycti import OpenCTIConnectorHelper
 
 if __name__ == "__main__":
-    greyNoiseInstance = GreyNoiseConnector()
-    greyNoiseInstance.start()
+    """
+    Entry point of the script
+
+    - traceback.print_exc(): This function prints the traceback of the exception to the standard error (stderr).
+    The traceback includes information about the point in the program where the exception occurred,
+    which is very useful for debugging purposes.
+    - exit(1): effective way to terminate a Python program when an error is encountered.
+    It signals to the operating system and any calling processes that the program did not complete successfully.
+    """
+    try:
+        settings = ConnectorSettings()
+        helper = OpenCTIConnectorHelper(
+            config=settings.to_helper_config(), playbook_compatible=True
+        )
+
+        connector = GreyNoiseConnector(config=settings, helper=helper)
+        connector.start()
+    except Exception:
+        traceback.print_exc()
+        exit(1)
