+++ internal-enrichment/greynoise/src/requirements.txt
@@ -1,3 +1,5 @@
 pycti==7.260710.0
 pycountry==24.6.1
 greynoise==3.0.1
+pydantic >=2.8.2, <3
+connectors-sdk @ git+https://github.com/OpenCTI-Platform/connectors.git@master#subdirectory=connectors-sdk
