+++ internal-enrichment/greynoise/tests/test-requirements.txt
@@ -1,0 +1,2 @@
+-r ../src/requirements.txt
+pytest==8.4.2
